Cross-validation for penalized discrete survival model (without provider information)

Performs k-fold cross validation for penalized discrete survival model (without provider information) over a grid of values of regularization parameter lambda.

cv.DiscSurv(
  data,
  Event.char,
  Z.char,
  Time.char,
  penalize.x = rep(1, length(Z.char)),
  ...,
  nfolds = 10,
  seed,
  fold,
  trace.cv = FALSE
)

Arguments

data

an dataframe or list object that contains the variables in the model.

Event.char

name of the event indicator in data as a character string. Event indicator should be a binary variable with 1 indicating that the event has occurred and 0 indicating (right) censoring.

Z.char

names of covariates in data as vector of character strings.

Time.char

name of the follow up time in data as a character string.

penalize.x

a vector indicates whether the corresponding covariate will be penalized, as in pp.DiscSurv function.

...

extra arguments to be passed to function.

nfolds

the number of cross-validation folds. Default is 10.

seed

the seed of the random number generator in order to obtain reproducible results.

fold

a vector that specifies the fold that observations belongs to. By default the observations are randomly assigned.

trace.cv

cv.DiscSurv will provide user with the progress of cross validation if trace.cv = TRUE. Default is FALSE.

Value

An object with S3 class cv.DiscSurv.

cve

the error for each value of lambda, averaged across the cross-validation folds.

cvse

the estimated standard error associated with each value of for cve.

lambda

the sequence of regularization parameter values along which the cross-validation error was calculated.

fit

the fitted pp.DiscSurv object for the whole data.

fold

the fold assignments for cross-validation for each observation

min

the index of lambda corresponding to lambda.min.

lambda.min

the value of lambda with the minimum cross-validation error.

References

K. He, J. Kalbfleisch, Y. Li, and et al. (2013) Evaluating hospital readmission rates in dialysis facilities; adjusting for hospital effects. Lifetime Data Analysis, 19: 490-512.

Examples

data(Surv_Data)
data <- Surv_Data$data
Event.char <- Surv_Data$Event.char
Z.char <- Surv_Data$Z.char
Time.char <- Surv_Data$Time.char
cv.fit <- cv.DiscSurv(data, Event.char, Z.char, Time.char, nfolds = 10, trace.cv = T)
#> Starting CV fold #1...
#> Starting CV fold #2...
#> Starting CV fold #3...
#> Starting CV fold #4...
#> Starting CV fold #5...
#> Starting CV fold #6...
#> Starting CV fold #7...
#> Starting CV fold #8...
#> Starting CV fold #9...
#> Starting CV fold #10...
cv.fit$cve
#>   [1] 0.6147027 0.5836890 0.5567544 0.5339292 0.5141126 0.4955502 0.4788320
#>   [8] 0.4642467 0.4514836 0.4402719 0.4304003 0.4217020 0.4140161 0.4072112
#>  [15] 0.4012332 0.3959862 0.3913945 0.3873817 0.3838861 0.3808499 0.3782231
#>  [22] 0.3759505 0.3740005 0.3723279 0.3708975 0.3696816 0.3686543 0.3677870
#>  [29] 0.3670620 0.3664553 0.3659529 0.3655593 0.3653111 0.3650681 0.3639319
#>  [36] 0.3625787 0.3614188 0.3604495 0.3596519 0.3589597 0.3583682 0.3578583
#>  [43] 0.3573946 0.3570125 0.3566985 0.3564433 0.3562350 0.3560667 0.3559306
#>  [50] 0.3558233 0.3557380 0.3556700 0.3556160 0.3555784 0.3555501 0.3555277
#>  [57] 0.3555118 0.3555036 0.3554982 0.3554952 0.3554970 0.3554992 0.3555010
#>  [64] 0.3555056 0.3555114 0.3555181 0.3555241 0.3555290 0.3555371 0.3555430
#>  [71] 0.3555483 0.3555562 0.3555586 0.3555634 0.3555697 0.3555716 0.3555793
#>  [78] 0.3555835 0.3555837 0.3555908 0.3555943 0.3555938 0.3555991 0.3555991
#>  [85] 0.3556040 0.3556074 0.3556074 0.3556090 0.3556111 0.3556124 0.3556123
#>  [92] 0.3556150 0.3556166 0.3556172 0.3556187 0.3556215 0.3556224 0.3556219
#>  [99] 0.3556239 0.3556233
cv.fit$lambda.min
#> [1] 0.001760446