Cross-validation for pp.DiscSurv

Performs k-fold cross validation for penalized regression models over a grid of values of regularization parameter lambda.

cv.pp.DiscSurv(
  data,
  Event.char,
  prov.char,
  Z.char,
  Time.char,
  penalize.x = rep(1, length(Z.char)),
  ...,
  nfolds = 10,
  seed,
  fold,
  trace.cv = FALSE
)

Arguments

data

an `dataframe` or `list` object that contains the variables in the model.

Event.char

name of the event indicator in `data` as a character string.

prov.char

name of provider IDs variable in `data` as a character string.

Z.char

names of covariates in `data` as vector of character strings.

Time.char

name of the observation time in `data` as a character string.

penalize.x

a vector indicates whether the corresponding covariate will be penalized, as in pp.DiscSurv function.

...

extra arguments to be passed to function.

nfolds

the number of cross-validation folds. Default is 10.

seed

the seed of the random number generator in order to obtain reproducible results.

fold

a vector that specifies the fold that observations belongs to. By default the observations are randomly assigned.

trace.cv

cv.pp.DiscSurv will provide user with the progress of cross validation if `trace.cv = TRUE`. Default is FALSE.

Value

An object with S3 class cv.pp.DiscSurv.

cve

the error for each value of lambda, averaged across the cross-validation folds.

cvse

the estimated standard error associated with each value of for cve.

lambda

the sequence of regularization parameter values along which the cross-validation error was calculated.

fit

the fitted pp.DiscSurv object for the whole data.

fold

the fold assignments for cross-validation for each observation

min

the index of lambda corresponding to lambda.min.

lambda.min

the value of lambda with the minimum cross-validation error.

References

K. He, J. Kalbfleisch, Y. Li, and et al. (2013) Evaluating hospital readmission rates in dialysis facilities; adjusting for hospital effects. Lifetime Data Analysis, 19: 490-512.

Examples

data(Surv_Data)
data <- Surv_Data$data
Event.char <- Surv_Data$Event.char
prov.char <- Surv_Data$prov.char
Z.char <- Surv_Data$Z.char
Time.char <- Surv_Data$Time.char
cv.fit <- cv.pp.DiscSurv(data, Event.char, prov.char, Z.char, Time.char, nfolds = 10, trace.cv = T)
#> Starting CV fold #1...
#> Starting CV fold #2...
#> Starting CV fold #3...
#> Starting CV fold #4...
#> Starting CV fold #5...
#> Starting CV fold #6...
#> Starting CV fold #7...
#> Starting CV fold #8...
#> Starting CV fold #9...
#> Starting CV fold #10...
cv.fit$cve
#>   [1] 0.5570408 0.4654882 0.4541073 0.4445862 0.4365787 0.4298319 0.4241022
#>   [8] 0.4184227 0.4124547 0.4068606 0.4020815 0.3980066 0.3945837 0.3916889
#>  [15] 0.3892216 0.3870409 0.3851005 0.3834329 0.3820309 0.3808578 0.3798811
#>  [22] 0.3790833 0.3783516 0.3772634 0.3757653 0.3745352 0.3735369 0.3727326
#>  [29] 0.3720931 0.3715562 0.3711336 0.3707951 0.3704853 0.3701942 0.3699301
#>  [36] 0.3697237 0.3695625 0.3694400 0.3693438 0.3692745 0.3692243 0.3691933
#>  [43] 0.3691710 0.3691590 0.3691595 0.3691665 0.3691698 0.3691822 0.3691964
#>  [50] 0.3692101 0.3692293 0.3692444 0.3692636 0.3692804 0.3692978 0.3693137
#>  [57] 0.3693293 0.3693459 0.3693592 0.3693735 0.3693850 0.3693966 0.3694097
#>  [64] 0.3694187 0.3694291 0.3694386 0.3694452 0.3694521 0.3694633 0.3694685
#>  [71] 0.3694735 0.3694798 0.3694844 0.3694903 0.3694920 0.3694959 0.3694994
#>  [78] 0.3695020 0.3695079 0.3695089 0.3695110 0.3695127 0.3695174 0.3695185
#>  [85] 0.3695202 0.3695222 0.3695233 0.3695230 0.3695266 0.3695280 0.3695285
#>  [92] 0.3695299 0.3695304 0.3695313 0.3695322 0.3695328 0.3695330 0.3695342
#>  [99] 0.3695348 0.3695344
cv.fit$lambda.min
#> [1] 0.002930123