predictions of a ppLasso or gr_ppLasso object

return the model predictions of a ppLasso or gr_ppLasso object

# S3 method for ppLasso
predict(
  fit,
  data,
  Z.char,
  prov.char,
  lambda,
  which = 1:length(fit$lambda),
  type = c("response", "class", "vars", "nvars"),
  ...
)

# S3 method for gr_ppLasso
predict(
  fit,
  data,
  Z.char,
  prov.char,
  lambda,
  which = 1:length(fit$lambda),
  type = c("response", "class", "vars", "groups", "nvars", "ngroups", "beta.norm"),
  ...
)

Arguments

fit

a ppLasso or gr_ppLasso or ppDiscSurv object.

data

an `dataframe` or `list` object that contains the variables for prediction.

Z.char

names of covariates in `data` as vector of character strings.

prov.char

name of provider IDs variable in `data` as a character string.

lambda

values of the regularization parameter lambda at which predictions are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.

which

indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.

type

type of prediction: response provides the fitted value; class returns the binomial outcome with the largest probability; vars returns the indices for the non-zero coefficients; nvars returns the number of non-zero coefficients; groups returns the indices for the non-zero groups; ngroups returns the number of non-zero coefficients; beta.norm returns L2 norm of the coefficients in each group

...

Y.char

name of the response variable in `data` as a character string.

Examples

data(GLM_Data)
data <- GLM_Data$data
Y.char <- GLM_Data$Y.char
prov.char <- GLM_Data$prov.char
Z.char <- GLM_Data$Z.char
fit <- pp.lasso(data, Y.char, Z.char, prov.char)
predict(fit, data, Z.char, prov.char, lambda = fit$lambda, type = "response")[1:10, 1:5]
#>          0.132    0.1202    0.1096    0.0998     0.091
#> 733  0.2874999 0.2996927 0.3106971 0.3206762 0.3297643
#> 853  0.2051302 0.2112109 0.2163924 0.2208163 0.2246015
#> 813  0.2051302 0.2191951 0.2321638 0.2441339 0.2551989
#> 576  0.3947391 0.4208778 0.4449423 0.4671087 0.4875485
#> 968  0.2345694 0.2340619 0.2330311 0.2316163 0.2299295
#> 1382 0.3188411 0.3744516 0.4280555 0.4786478 0.5256182
#> 1072 0.3924033 0.3836025 0.3752418 0.3673071 0.3597819
#> 330  0.3658512 0.3870533 0.4065558 0.4245437 0.4411742
#> 204  0.2298886 0.2256943 0.2215601 0.2175107 0.2135686
#> 614  0.3947391 0.4374814 0.4772352 0.5139266 0.5476058
predict(fit, data, Z.char, prov.char, lambda = 0.001, type = "class")[1:10]
#>  [1] 1 0 0 1 0 1 0 1 0 1
predict(fit, data, Z.char, prov.char, lambda = 0.04, type = "vars")
#> Z1 Z3 Z4 
#>  1  3  4 

data(GLM_Data)
data <- GLM_Data$data
Y.char <- GLM_Data$Y.char
prov.char <- GLM_Data$prov.char
Z.char <- GLM_Data$Z.char
group <- GLM_Data$group
fit <- grp.lasso(data, Y.char, Z.char, prov.char, group = group)
predict(fit, GLM_Data, Z.char, prov.char, lambda = fit$lambda, type = "response")[1:10, 1:5]
#> Error in data[, prov.char, drop = F]: incorrect number of dimensions
predict(fit, GLM_Data, Z.char, prov.char, lambda = 0.001, type = "class")[1:10]
#> Error in data[, prov.char, drop = F]: incorrect number of dimensions
predict(fit, GLM_Data, Z.char, prov.char, lambda = 0.04, type = "vars")
#> Z1 Z2 Z3 Z4 
#>  1  2  3  4 
predict(fit, GLM_Data, Z.char, prov.char, lambda = 0.04, type = "groups")
#> [1] 1 2
#> Levels: 1 2 3