Predictions of a strat_cox object

Return the model predictions of a strat_cox object

# S3 method for strat_cox
predict(
  fit,
  data,
  Event.char,
  prov.char,
  Z.char,
  Time.char,
  lambda,
  which = 1:length(fit$lambda),
  type = c("link", "response", "vars", "nvars", "groups", "ngroups"),
  ...
)

Arguments

fit

a strat_cox object.

data

an dataframe or list object that contains the variables for prediction.

Event.char

name of the event indicator in data as a character string.

prov.char

name of provider IDs variable in data as a character string.

Z.char

names of covariates in data as vector of character strings.

Time.char

name of the observation time in data as a character string.

lambda

values of the regularization parameter lambda at which predictions are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.

which

indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.

type

type of prediction:

link: linear predictors

response: risk (i.e., exp(link))

vars: the indices for the non-zero coefficients

nvars: the number of non-zero coefficients

groups: the indices for the non-zero groups

ngroups: the number of non-zero coefficients

...

Examples

data(Cox_Data)
data <- Cox_Data$data
Event.char <- Cox_Data$Event.char
prov.char <- Cox_Data$prov.char
Z.char <- Cox_Data$Z.char
Time.char <- Cox_Data$Time.char
fit <- Strat.cox(data, Event.char, prov.char, Z.char, Time.char, group = c(1, 2, 2, 3, 3))
predict(fit, data, Event.char, prov.char, Z.char, Time.char, lambda = fit$lambda, type = "response")[1:5, 1:5]
#>      0.0394    0.0359    0.0327    0.0298    0.0272
#> [1,]      1 0.9134386 0.8391602 0.7768358 0.7240672
#> [2,]      1 0.9290607 0.8671789 0.8144614 0.7692052
#> [3,]      1 0.8660430 0.7568822 0.6695592 0.5987700
#> [4,]      1 0.9381111 0.8836146 0.8367839 0.7962623
#> [5,]      1 0.8574967 0.7424831 0.6512926 0.5779641